requirements.txt
<<<<<<< HEAD
Flask
gunicorn
werkzeug
google-generativeai
pillow
python-docx
odfdo
firebase-admin
google-genai
markdown
PyPDF2
pdf2image
python-dotenv
google-cloud-secret-manager
=======
flask
python-docx
odfdo
markdown
requests
werkzeug
google-generativeai
google-cloud-storage
google-cloud-vision
google-cloud-documentai
google-auth
dotenv
gunicorn
>>>>>>> 3044c46a7f485dca27def97e2673f28b3bc69065
